ai_service/requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
pillow
ftfy
regex
git+https://github.com/openai/CLIP.git
python-multipart